Vendor Import v4.3.12
[BUGFIX] Untrusted GP data is unserialized in wizard_colorpicker.php and view_help.php (Christian Kuhn)
[BUGFIX] Unprivileged backend user can read arbitrarily from database (Helmut Hummel)
[BUGFIX] XSS in TYPO3 core when using typolink.parameter JS-Popup Window (Marco Bresch)
[BUGFIX] XSS in tcemain flash message (Georg Ringer)
[BUGFIX] XSS in Recycler (Steffen Gebert)
[BUGFIX] XSS in browse_links (Georg Ringer)
[BUGFIX] t3lib_div::removeXSS does not remove some XSS (Jigal van Hemert)
[BUGFIX] XSS in admin panel (Georg Ringer)
[BUGFIX] The delay if wrong credentials are provided during backend login can be circumvented (Helmut Hummel)
[BUGFIX] Information disclosure during backend login (Helmut Hummel)
[BUGFIX] Value 0 in integer fields is not shown in List module anymore (Benjamin Mack)
[BUGFIX] Date of 1970-01-01 is shown in unset date fields (Steffen Gebert)
[BUGFIX] FE-Error "NO entry in the $TCA-array..." (Nicole Cordes)
[TASK] Integrate external projects Fluid and Extbase as Git submodules (Oliver Hader)
[TASK] Use external Core Project DBAL as Git submodule (Oliver Hader)
[TASK] Use external Core Project version as Git submodule (Oliver Hader)
[TASK] Use external Core Project version as Git submodule (Oliver Hader)
[TASK] Remove file SVNreadme.txt (Steffen Gebert)
[BUGFIX] Flexformfield "Forgot Password Message" in felogin does not work (Steffen Mueller)
[FEATURE] add .gitignore for common file exclusions (Peter Niederlag(office))

+++ t3lib/class.t3lib_lock.php
@@ -27,7 +27,7 @@
 /**
  * Class for providing locking features in TYPO3
  *
- * $Id: class.t3lib_lock.php 9863 2010-12-20 19:42:03Z tolleiv $
+ * $Id$
  *
  * @author	Michael Stucki <michael@typo3.org>
  */
@@ -214,8 +214,10 @@ class t3lib_lock {
 		$success = TRUE;
 		switch ($this->method) {
 			case 'simple':
-				if (unlink($this->resource) == FALSE) {
-					$success = FALSE;
+				if (t3lib_div::isAllowedAbsPath($this->resource) && t3lib_div::isFirstPartOfStr($this->resource, PATH_site . 'typo3temp/locks/')) {
+					if (unlink($this->resource) == FALSE) {
+						$success = FALSE;
+					}
 				}
 			break;
 			case 'flock':
@@ -223,7 +225,9 @@ class t3lib_lock {
 					$success = FALSE;
 				}
 				fclose($this->filepointer);
-				unlink($this->resource);
+				if (t3lib_div::isAllowedAbsPath($this->resource) && t3lib_div::isFirstPartOfStr($this->resource, PATH_site . 'typo3temp/locks/')) {
+					unlink($this->resource);
+				}
 			break;
 			case 'semaphore':
 				if (@sem_release($this->resource)) {
